+++ packages/react-native-renderer/src/ReactNativeRenderer.js
@@ -12,6 +12,7 @@ import type {ReactNodeList} from 'shared/ReactTypes';
 
 import './ReactNativeInjection';
 
+import ReactFiberReconciler from 'react-reconciler';
 import * as ReactPortal from 'shared/ReactPortal';
 import * as ReactGenericBatching from 'events/ReactGenericBatching';
 import ReactVersion from 'shared/ReactVersion';
@@ -20,15 +21,17 @@ import UIManager from 'UIManager';
 
 import {getStackAddendumByWorkInProgressFiber} from 'shared/ReactFiberComponentTreeHook';
 
+import ReactNativeHostConfig from './ReactNativeHostConfig';
 import NativeMethodsMixin from './NativeMethodsMixin';
 import ReactNativeComponent from './ReactNativeComponent';
 import * as ReactNativeComponentTree from './ReactNativeComponentTree';
-import ReactNativeFiberRenderer from './ReactNativeFiberRenderer';
 import {getInspectorDataForViewTag} from './ReactNativeFiberInspector';
 
 import {ReactCurrentOwner} from 'shared/ReactGlobalSharedState';
 import getComponentName from 'shared/getComponentName';
 import warning from 'fbjs/lib/warning';
+
+const ReactNativeFiberRenderer = ReactFiberReconciler(ReactNativeHostConfig);
 
 const findHostInstance = ReactNativeFiberRenderer.findHostInstance;
